LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/foreign-fetch-cors.https-expected.txt

CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Afalse%2C%22with_aceheaders%22%3Afalse%2C%22with_acaorigin%22%3Afalse%7D. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Afalse%2C%22with_aceheaders%22%3Afalse%2C%22with_acaorigin%22%3Afalse%7D due to access control checks.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Afalse%2C%22with_aceheaders%22%3Afalse%2C%22with_acaorigin%22%3Afalse%7D. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Afalse%2C%22with_aceheaders%22%3Afalse%2C%22with_acaorigin%22%3Afalse%7D. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Afalse%2C%22with_aceheaders%22%3Afalse%2C%22with_acaorigin%22%3Afalse%7D. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Afalse%2C%22with_aceheaders%22%3Afalse%2C%22with_acaorigin%22%3Afalse%7D due to access control checks.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Afalse%2C%22with_aceheaders%22%3Atrue%2C%22with_acaorigin%22%3Atrue%7D. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Afalse%2C%22with_aceheaders%22%3Atrue%2C%22with_acaorigin%22%3Atrue%7D due to access control checks.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Afalse%2C%22with_aceheaders%22%3Atrue%2C%22with_acaorigin%22%3Atrue%7D. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Afalse%2C%22with_aceheaders%22%3Atrue%2C%22with_acaorigin%22%3Atrue%7D. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Afalse%2C%22with_aceheaders%22%3Atrue%2C%22with_acaorigin%22%3Atrue%7D. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Afalse%2C%22with_aceheaders%22%3Atrue%2C%22with_acaorigin%22%3Atrue%7D due to access control checks.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Atrue%2C%22with_aceheaders%22%3Afalse%2C%22with_acaorigin%22%3Afalse%7D. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Atrue%2C%22with_aceheaders%22%3Afalse%2C%22with_acaorigin%22%3Afalse%7D due to access control checks.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Atrue%2C%22with_aceheaders%22%3Atrue%2C%22with_acaorigin%22%3Atrue%7D. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Atrue%2C%22with_aceheaders%22%3Atrue%2C%22with_acaorigin%22%3Atrue%7D due to access control checks.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Atrue%2C%22with_aceheaders%22%3Atrue%2C%22with_acaorigin%22%3Atrue%7D. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Atrue%2C%22with_aceheaders%22%3Atrue%2C%22with_acaorigin%22%3Atrue%7D. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Atrue%2C%22with_aceheaders%22%3Atrue%2C%22with_acaorigin%22%3Atrue%7D. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Atrue%2C%22with_aceheaders%22%3Atrue%2C%22with_acaorigin%22%3Atrue%7D due to access control checks.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Atrue%2C%22with_aceheaders%22%3Afalse%2C%22with_acaorigin%22%3Atrue%7D. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Atrue%2C%22with_aceheaders%22%3Afalse%2C%22with_acaorigin%22%3Atrue%7D due to access control checks.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Atrue%2C%22with_aceheaders%22%3Afalse%2C%22with_acaorigin%22%3Atrue%7D. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/simple.txt?%7B%22cross_origin%22%3Atrue%2C%22with_aceheaders%22%3Afalse%2C%22with_acaorigin%22%3Atrue%7D. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
  

PASS Same origin fetch without CORS headers, not exposed 
FAIL Same origin fetch without CORS headers, only origin exposed promise_test: Unhandled rejection with value: object "TypeError: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin."
FAIL Same origin fetch without CORS headers, headers and origin exposed promise_test: Unhandled rejection with value: object "TypeError: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin."
FAIL Same origin fetch without CORS headers, exposed to wrong origin assert_unreached: Should have rejected: undefined Reached unreachable code
PASS Same origin fetch with CORS headers, not exposed 
FAIL Same origin fetch with CORS headers, only origin exposed promise_test: Unhandled rejection with value: object "TypeError: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin."
FAIL Same origin fetch with CORS headers, headers and origin exposed promise_test: Unhandled rejection with value: object "TypeError: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin."
FAIL Same origin fetch with CORS headers, exposed to wrong origin assert_unreached: Should have rejected: undefined Reached unreachable code
PASS Cross origin fetch without CORS headers, not exposed 
PASS Cross origin fetch with ACEHeaders header, not exposed 
FAIL Cross origin fetch with ACEHeaders header, only origin exposed promise_test: Unhandled rejection with value: object "TypeError: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin."
FAIL Cross origin fetch with ACEHeaders header, headers and origin exposed promise_test: Unhandled rejection with value: object "TypeError: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin."
FAIL Cross origin fetch with ACEHeaders header, exposed to wrong origin assert_unreached: Should have rejected: undefined Reached unreachable code
PASS Cross origin fetch without ACEHeaders header, not exposed 
FAIL Cross origin fetch without ACEHeaders header, only origin exposed promise_test: Unhandled rejection with value: object "TypeError: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin."
FAIL Cross origin fetch without ACEHeaders header, headers and origin exposed promise_test: Unhandled rejection with value: object "TypeError: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin."
